Add actions support

+++ apps/hyperdb/src/hyperdb/subscribable-db.ts
@@ -1,34 +1,29 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import type {
-  DB,
-  ExtractSchema,
-  Row,
-  ScanOptions,
-  TableDefinition,
-} from "./db";
+import type { DB, HyperDB, Row, SelectOptions, TupleScanOptions } from "./db";
+import type { ExtractIndexes, ExtractSchema, TableDefinition } from "./table";
 
 export type InsertOp = {
   type: "insert";
-  table: TableDefinition<any>;
+  table: TableDefinition;
   newValue: Row;
 };
 
 export type UpdateOp = {
   type: "update";
-  table: TableDefinition<any>;
+  table: TableDefinition;
   oldValue: Row;
   newValue: Row;
 };
 
 export type DeleteOp = {
   type: "delete";
-  table: TableDefinition<any>;
+  table: TableDefinition;
   oldValue: Row;
 };
 
 export type Op = InsertOp | UpdateOp | DeleteOp;
 
-export class SubscribableDB {
+export class SubscribableDB implements HyperDB {
   private subscribers: ((op: Op[]) => void)[] = [];
   private db: DB;
 
@@ -44,35 +39,24 @@ export class SubscribableDB {
     };
   }
 
-  *hashScan<TTable extends TableDefinition<any>>(
+  *intervalScan<
+    TTable extends TableDefinition<any, any>,
+    K extends keyof ExtractIndexes<TTable>,
+  >(
     table: TTable,
-    indexName: string,
-    ids: string[],
+    indexName: K,
+    options: TupleScanOptions[],
+    selectOptions?: SelectOptions,
   ): Generator<ExtractSchema<TTable>> {
-    for (const data of this.db.hashScan(table, indexName, ids)) {
-      yield data;
+    if (options && options.length === 0) {
+      throw new Error("scan options must be provided");
     }
-  }
 
-  *intervalScan<TTable extends TableDefinition<any>>(
-    table: TTable,
-    indexName: keyof TTable["indexes"],
-    options?: ScanOptions,
-  ): Generator<ExtractSchema<TTable>> {
-    for (const data of this.db.intervalScan(table, indexName, options)) {
-      yield data;
-    }
-  }
-
-  async *asyncScan<TTable extends TableDefinition<any>>(
-    table: TTable,
-    indexName: keyof TTable["indexes"],
-    options?: ScanOptions,
-  ): AsyncGenerator<ExtractSchema<TTable>> {
-    for await (const data of this.db.asyncIntervalScan(
+    for (const data of this.db.intervalScan(
       table,
       indexName,
       options,
+      selectOptions,
     )) {
       yield data;
     }
@@ -82,6 +66,8 @@ export class SubscribableDB {
     table: TTable,
     records: ExtractSchema<TTable>[],
   ) {
+    if (records.length === 0) return;
+
     this.db.insert(table, records);
     const ops = records.map(
       (record): Op => ({
@@ -97,12 +83,14 @@ export class SubscribableDB {
     table: TTable,
     records: ExtractSchema<TTable>[],
   ) {
+    if (records.length === 0) return;
+
     const previousRecords = new Map<string, Row>();
 
-    for (const oldRecord of this.db.hashScan(
+    for (const oldRecord of this.db.intervalScan(
       table,
       table.idIndexName,
-      records.map((r) => r.id),
+      records.map((r) => ({ gte: [r.id], lte: [r.id] })),
     )) {
       previousRecords.set(oldRecord.id, oldRecord);
     }
@@ -110,7 +98,7 @@ export class SubscribableDB {
     for (const record of records) {
       if (!previousRecords.has(record.id)) {
         throw new Error(
-          `Failed to update record, no previous record found for ${table.name}=${record.id}`,
+          `Failed to update record, no previous record found for ${table.tableName}=${record.id}`,
         );
       }
     }
@@ -130,8 +118,14 @@ export class SubscribableDB {
   }
 
   delete<TTable extends TableDefinition<any>>(table: TTable, ids: string[]) {
+    if (ids.length === 0) return;
+
     const opsToNotify: DeleteOp[] = [];
-    for (const oldRecord of this.db.hashScan(table, table.idIndexName, ids)) {
+    for (const oldRecord of this.db.intervalScan(
+      table,
+      table.idIndexName,
+      ids.map((id) => ({ gte: [id], lte: [id] })),
+    )) {
       opsToNotify.push({ type: "delete", table, oldValue: oldRecord });
     }
 
@@ -139,4 +133,27 @@ export class SubscribableDB {
 
     this.subscribers.forEach((s) => s(opsToNotify));
   }
+
+  // *hashScan<TTable extends TableDefinition<any>>(
+  //   table: TTable,
+  //   indexName: string,
+  //   ids: string[],
+  // ): Generator<ExtractSchema<TTable>> {
+  //   for (const data of this.db.hashScan(table, indexName, ids)) {
+  //     yield data;
+  //   }
+  // }
+  // async *asyncScan<TTable extends TableDefinition<any>>(
+  //   table: TTable,
+  //   indexName: keyof TTable["indexes"],
+  //   options?: ScanOptions,
+  // ): AsyncGenerator<ExtractSchema<TTable>> {
+  //   for await (const data of this.db.asyncIntervalScan(
+  //     table,
+  //     indexName,
+  //     options,
+  //   )) {
+  //     yield data;
+  //   }
+  // }
 }
